cleanup: improve docs
